Enable Web snapshot tests to include CSS
WKWebView has prevention which stops it from being able to access file URLs, especially those which are outside the children of the HTML file itself.

When loading a HTML string this means it can't access file URLs due to the sandboxing of tests and the web view.

Apple seemed to add support for this by adding a "loadFileURL allowingReadAccessTo" method. In order to use this we must first save the HTML as a file within that directory (I call it "index.html" though this could change)

It's important to know that due to the way Apple loads relative resources, the CSS file (or any resource) can not have a leading slash.