Separate search query from search term
This adds two new values to the search response. A "term" which is a subset of the query minus the filters, and "filters" which is currently an array of strings but this is likely to change in a future commit (just an empty array throughout).

The search term is different to the search query as with a query of "test stars:>500 framework" the term is simply "test framework" and the filters is ["stars:>500"].

This is because in the UI for the search results, we want to show the term separate to the filters. However, we still need the raw queries for some parts of the UI (for example the user-input search box).