Enable GZIP compression for all responses
As per the documentation (https://docs.vapor.codes/4.0/server/#response-compression) this is turned off by default, but does provide (in testing) a decrease in response times and data sizes. Hopefully this will help reduce our bandwidth, reducing costs, as well as making the experience better for our users.

We will want to closely monitor these changes, it may be hard to evaluate perfectly in staging due to the low usage.